+++ gpsd/timehint.c
@@ -472,10 +472,11 @@ void ntpshm_link_activate(struct gps_device_t *session)
 {
     struct gps_context_t *context = session->context;
 
+    // cast for 32-bit intptr_t
     GPSD_LOG(LOG_PROG, &context->errout,
              "NTP:SHM: ntpshm_link_activate(%s), sourcetype %d fd %ld\n",
              session->gpsdata.dev.path, session->sourcetype,
-             session->gpsdata.gps_fd);
+             (long)session->gpsdata.gps_fd);
 
     /* Don't talk to NTP when we're:
      *   reading from a file
